로그인 상태 수정
