+++ client/src/components/Button.jsx
@@ -1,6 +1,7 @@
+import React from 'react';
 import styled from 'styled-components';
 
-function Button({ children, src }) {
+const Button = ({ children, src }) => {
   // nav바 상단 카드만들기, 닉네임 부분 
 
   return (
