+++ client/src/components/Jumbotron.jsx
@@ -1,7 +1,8 @@
+import React from 'react';
 import styled from 'styled-components';
 import SearchBar from './SearchBar';
 
-function Jumbotron() {
+const Jumbotron = () => {
   return (
     <StyledJumbotron>
       <Banner>
