+++ client/src/components/MycardList.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const MycardList = () => {
     return (
-            <h2>케케케</h2>
+            <div></div>
     );
 };
