+++ client/src/containers/SettingContainer.jsx
@@ -1,6 +1,7 @@
+import React from 'react';
 import styled from 'styled-components';
 
-function SettingContainer() {
+const SettingContainer = () => {
   return (
     <SettingWrap>
       <SettingLeft>
